# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fake Store API",
        "description": "This is the API documentation for my Laravel project.",
        "contact": {
            "email": "your-email@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get paginated list of products",
                "operationId": "796ff0bdf4018b2d3636a3f65e72e689",
                "parameters": [
                    {
                        "name": "_pageLimit",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "_pageSize",
                        "in": "query",
                        "description": "Current page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with paginated products"
                    }
                }
            }
        },
        "/api/product/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get product by ID",
                "operationId": "0847d7d2b6de240f2532a6a0a28060a8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product found"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
        "/api/product/create": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create a new product",
                "operationId": "3f93e9ef90a05362bc5d70f7c0721087",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_name",
                                    "price",
                                    "cost",
                                    "category_id"
                                ],
                                "properties": {
                                    "product_name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "cost": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/product/update": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Update product",
                "operationId": "cc5d4dca3996bc7a9f428b3e83392dec",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "product_name",
                                    "price",
                                    "cost",
                                    "category_id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "product_name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "cost": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/product/delete/force": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Permanently delete a soft-deleted product",
                "operationId": "bfec7733e6eab95caf461d63ba7a2d63",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product permanently deleted"
                    },
                    "404": {
                        "description": "Product not found in trash"
                    }
                }
            }
        },
        "/api/product/delete/soft": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Soft delete a product",
                "operationId": "f1505cfa5fb691572304e4bad6636b1d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product soft deleted"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
        "/api/product/restore/{id}": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Restore a soft-deleted product",
                "operationId": "fa1a19f8528ce6e6bb37d3122ec79da1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product restored successfully"
                    },
                    "404": {
                        "description": "Product not found in trash"
                    }
                }
            }
        },
...
    },
...
}